requirements.txt
requests==2.31.0
faker==24.9.0
azure-storage-blob==12.19.0
python-dotenv==1.0.1
azure-functions
azure-storage-blob
azure-cosmos
azure-servicebus
